# apps/backend/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API documentation for IgniticAI backend",
        "title": "IgniticAI API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
...
  "/api/v1/organizations/{id}/business-profile": {
  "post": {
    "summary": "Create or update organization business profile",
    "description": "Creates a new organization business profile if it doesn’t exist, otherwise updates the existing profile. Only organization admins can perform this action.",
    "tags": ["organization-business-profile"],
    "consumes": ["application/json"],
    "produces": ["application/json"],
    "security": [{ "BearerAuth": [] }],
    "parameters": [
      {
        "name": "id",
        "in": "path",
        "required": true,
        "type": "string",
        "description": "Organization ID"
      },
      {
        "name": "profile",
        "in": "body",
        "required": true,
        "schema": {
          "type": "object",
          "properties": {
            "business_hours": { "type": "string" },
            "primary_markets": { "type": "array", "items": { "type": "string" } },
            "default_currency": { "type": "string" },
            "supported_languages": { "type": "array", "items": { "type": "string" } },
            "support_email": { "type": "string" },
            "support_channels": { "type": "array", "items": { "type": "string" } },
            "social_links": { "type": "object", "additionalProperties": { "type": "string" } },
            "fulfillment_method": { "type": "string" },
            "shipping_carriers": { "type": "array", "items": { "type": "string" } },
            "returns_policy_url": { "type": "string" },
            "payment_gateways": { "type": "array", "items": { "type": "string" } },
            "tax_identifiers": { "type": "object", "additionalProperties": { "type": "string" } },
            "primary_contacts": { "type": "array", "items": { "type": "object" } },
            "compliance_contacts": { "type": "array", "items": { "type": "object" } },
            "ecommerce_platforms": { "type": "array", "items": { "type": "object" } },
            "key_systems": { "type": "array", "items": { "type": "string" } },
            "holiday_blackout_dates": { "type": "array", "items": { "type": "string" } },
            "data_processing_addenda": { "type": "string" }
          }
        }
      }
    ],
    "responses": {
      "200": { "description": "Business profile created or updated successfully", "schema": { "type": "object" } },
      "400": { "description": "Invalid request or ID format", "schema": { "type": "object" } },
      "401": { "description": "Unauthorized", "schema": { "type": "object" } },
      "403": { "description": "Forbidden – Only admin can modify business profile", "schema": { "type": "object" } },
      "500": { "description": "Server error", "schema": { "type": "object" } }
    }
  },
  "get": {
    "summary": "Get organization business profile",
    "description": "Retrieves the business profile of an organization. Only organization admins can view it.",
    "tags": ["organization-business-profile"],
    "produces": ["application/json"],
    "security": [{ "BearerAuth": [] }],
    "parameters": [
      {
        "name": "id",
        "in": "path",
        "required": true,
        "type": "string",
        "description": "Organization ID"
      }
    ],
    "responses": {
      "200": { "description": "Organization business profile retrieved successfully", "schema": { "type": "object" } },
      "400": { "description": "Invalid request or ID format", "schema": { "type": "object" } },
      "401": { "description": "Unauthorized", "schema": { "type": "object" } },
      "403": { "description": "Forbidden – Only admin can view business profile", "schema": { "type": "object" } },
      "404": { "description": "Business profile not found", "schema": { "type": "object" } }
    }
  },
  "delete": {
    "summary": "Delete organization business profile",
    "description": "Deletes an organization's business profile. Only organization admins can perform this action.",
    "tags": ["organization-business-profile"],
    "produces": ["application/json"],
    "security": [{ "BearerAuth": [] }],
    "parameters": [
      {
        "name": "id",
        "in": "path",
        "required": true,
        "type": "string",
        "description": "Organization ID"
      }
    ],
    "responses": {
      "200": { "description": "Organization business profile deleted successfully", "schema": { "type": "object" } },
      "400": { "description": "Invalid request or ID format", "schema": { "type": "object" } },
      "401": { "description": "Unauthorized", "schema": { "type": "object" } },
      "403": { "description": "Forbidden – Only admin can delete business profile", "schema": { "type": "object" } },
      "500": { "description": "Server error", "schema": { "type": "object" } }
    }
  }
},
...
    }
}